stm32/rcc: Move CtlHsiClock() to common code.
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6200 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/include/arch/cm3/stm/stm32_clk.h
@@ -92,7 +92,7 @@ extern uint32_t STM_ClockGet(int idx);
 
 /* Control functions for the separate clocks */
 extern int CtlHseClock(int ena);
-extern int CtlHsiClock( uint8_t ena);
+extern int CtlHsiClock(int ena);
 extern int CtlPllClock( uint8_t ena);
 
 /* Selection functions for the clock sources */
